requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
line-bot-sdk==3.6.0
google-generativeai>=0.8.0
requests==2.31.0
python-dotenv==1.0.0
pillow==10.1.0
feedparser==6.0.10
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
lxml==4.9.3
markdown==3.5.1
python-multipart==0.0.6
aiofiles==23.2.1

# バッチ処理とOAuth用の追加依存関係
apscheduler==3.10.4
requests-oauthlib==1.3.1
xmltodict==0.13.0

# Google Photos API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0